requirements.txt
PyYAML==6.0.1
multicommand==1.0.0
termcolor==2.4.0
setuptools==69.1.0
deepmerge==1.1.1
deepdiff==6.7.1
urllib3==2.2.0